+++ CHANGELOG.md
@@ -99,6 +99,7 @@
 * nostr: add `notifications` field to NIP47 `GetInfoResponse` ([Yuki Kishimoto])
 * nostr: add `RelayMetadata::as_str` method ([Yuki Kishimoto])
 * nostr: add `nip42::is_valid_auth_event` function ([Yuki Kishimoto])
+* nostr: add `Tag::relays` constructor ([Yuki Kishimoto])
 * database: add `Events::force_insert` ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
 * pool: add `AdmitPolicy` trait ([Yuki Kishimoto])